tk_book_app_Shinevskiy/tk_book_app_Shinevskiy/requirements.txt
pillow>=10.0
